apply 4 space formatting, and use r-languageserver to format to standard
Language Server Protocols (LSPs) are incredible, not just for coding but
also for applying linting and formatting, and I recommend _everyone_ use
one. There must be a way to do it in RStudio, but I use Neovim (with the
R.nvim plugin) and using the LSP has changed my life.